Add removers for CurriculumInventoryReport
Also added SequenceBlocksEntity trait and applied it to academic levels,
report and session entities.
